backend/requirements_simple.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0

# AI models
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
diffusers==0.24.0
accelerate==0.25.0
safetensors==0.4.1

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Video processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# Async tasks
celery==5.3.4

# Utilities
requests==2.31.0
aiofiles==23.2.1 